requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
deprecation==2.1.0
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
img2pdf==0.4.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.3
packaging==23.1
pdf2jpg==1.1
pikepdf==7.2.0
Pillow==9.5.0
PyMuPDF==1.22.3
SQLAlchemy==2.0.16
typing_extensions==4.6.3
Werkzeug==2.3.6
